refactor(test): three more suites stop referencing MeshWeaver.AI
Hosting.Monolith.Test, NodeOperations.Test and Security.Test each drop their
ProjectReference to the AI engine. Everything AI-specific moves to MeshWeaver.AI.Test,
which travels to MeshWeaver.Plugins with the engine; what stays is core coverage that
merely used an AI type as a fixture.

  Hosting.Monolith.Test  4 AI tests moved (AgentChatClient, LanguageModelSyncedQuery,
                         AgentSyncedQueryFromHostedHub, AiSourcesInstallHook). The two
                         remaining `using MeshWeaver.AI;` were genuinely dead.
  NodeOperations.Test    MeshPluginAccessContextTest moved — it drives MeshPlugin tool
                         calls, which is AI's surface.
  Security.Test          ThreadAccessTest and ThreadStreamingIdentityTest moved. The Agent
                         case of StaticNamespacePolicyTests became AI.Test's
                         AgentNamespacePolicyTest; Doc and Role stay, because the contract
                         (static providers cap to read-only) is general and only the
                         INSTANCE belongs with the provider that emits it.

🚨 One re-point was wrong and the tests caught it. NodeCreationAccessTest's
CreateThread_UnderOwnUserNode_Succeeds was re-pointed to MarkdownNodeType as though the
node type were incidental. It is not: the test asserts MainNode points at the PARENT, which
is satellite wiring, and Thread is a satellite type while Markdown is not. It failed on
exactly that assertion. Substituting a type is only safe when the type is incidental to
what is asserted; when the test asserts a property the type CONFERS, substitution silently
changes the subject. That test now lives in AI.Test as ThreadSelfAccessCreationTest with
ThreadNodeType restored, while its sibling (creating under ANOTHER user's node must throw)
is genuinely type-agnostic and stays in core on Markdown.

RlsIntegrationTests' "Thread" fixture WAS incidental — row-level-security assertions say
nothing about the type — so it is re-pointed rather than moved.

Verified: all four projects build -c Release -p:CIRun=true -warnaserror at 0 Warning(s)/
0 Error(s); Security.Test's NodeCreationAccessTest passes 6/6, and the moved tests pass in
their new home (15 + 1).

No What's New entry: test-only refactor, no user-visible effect.

Refs #2276
