+++ test/MeshWeaver.AI.Test/MeshPluginAccessContextTest.cs
@@ -2,7 +2,6 @@ using System.Reactive.Threading.Tasks;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
-using MeshWeaver.AI;
 using MeshWeaver.AI.Persistence;
 using MeshWeaver.Graph;
 using MeshWeaver.Hosting.Monolith.TestBase;
@@ -16,7 +15,7 @@ using Microsoft.Extensions.AI;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 
-namespace MeshWeaver.NodeOperations.Test;
+namespace MeshWeaver.AI.Test;
 
 /// <summary>
 /// Tests that MeshPlugin tool calls restore the user's access context.
