+++ test/MeshWeaver.Hosting.Monolith.Test/CodeEditRecompileTest.cs
@@ -6,7 +6,6 @@ using System.Reactive.Threading.Tasks;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
-using MeshWeaver.AI;
 using MeshWeaver.Data;
 using MeshWeaver.Graph;
 using MeshWeaver.Graph.Configuration;
