+++ test/MeshWeaver.Security.Test/NodeCreationAccessTest.cs
@@ -4,7 +4,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Reactive.Threading.Tasks;
 using System.Reactive.Linq;
-using MeshWeaver.AI;
 using MeshWeaver.Graph.Configuration;
 using MeshWeaver.Hosting.Monolith.TestBase;
 using MeshWeaver.Hosting.Security;
@@ -31,8 +30,7 @@ public class NodeCreationAccessTest(ITestOutputHelper output) : MonolithMeshTest
     protected override MeshBuilder ConfigureMesh(MeshBuilder builder)
     {
         // ConfigureMeshBase adds RLS + graph (incl. Markdown)
-        return ConfigureMeshBase(builder)
-            .AddThreadType();
+        return ConfigureMeshBase(builder);
     }
 
     /// <summary>
@@ -244,48 +242,6 @@ public class NodeCreationAccessTest(ITestOutputHelper output) : MonolithMeshTest
     }
 
     /// <summary>
-    /// Tests that a user can create a Thread node under their own User node
-    /// via the self-access check (User/{ObjectId} scope).
-    /// This is the permission path used by the dashboard chat to create threads.
-    /// </summary>
-    [Fact(Timeout = 15000)]
-    public async Task CreateThread_UnderOwnUserNode_Succeeds()
-    {
-        // Arrange — log in as a user whose ObjectId matches a User node path
-        var accessService = Mesh.ServiceProvider.GetRequiredService<AccessService>();
-        var userId = "self-access-user";
-        var userContext = new AccessContext { ObjectId = userId, Name = "Self Access User" };
-        accessService.SetContext(userContext);
-        accessService.SetHostIdentity(userContext);
-
-        try
-        {
-            // Act — create a Thread under {userId} (self-access should grant permission).
-            // Post-v10: per-user partition lives at root namespace, so the user's
-            // own scope is just "{userId}" rather than the legacy "User/{userId}".
-            var threadPath = $"{userId}/TestThread_{Guid.NewGuid().AsString()}";
-            var threadNode = MeshNode.FromPath(threadPath) with
-            {
-                Name = "Test Chat Thread",
-                NodeType = ThreadNodeType.NodeType
-            };
-
-            var created = await NodeFactory.CreateNode(threadNode).Should().Emit();
-
-            // Assert
-            created.Should().NotBeNull("User should be able to create threads under their own User node");
-            created.State.Should().Be(MeshNodeState.Active);
-            created.Path.Should().Be(threadPath);
-            created.MainNode.Should().Be(userId, "Satellite thread MainNode should point to parent node");
-            Output.WriteLine($"Thread created successfully at: {created.Path}, MainNode: {created.MainNode}");
-        }
-        finally
-        {
-            TestUsers.DevLogin(Mesh);
-        }
-    }
-
-    /// <summary>
     /// Tests that a user CANNOT create a Thread under another user's node.
     /// </summary>
     [Fact(Timeout = 15000)]
@@ -299,19 +255,19 @@ public class NodeCreationAccessTest(ITestOutputHelper output) : MonolithMeshTest
 
         try
         {
-            // Act — try to create a thread under another user's node
+            // Act — try to create a node under another user's node
             var threadPath = "User/other-user/MaliciousThread";
             var threadNode = MeshNode.FromPath(threadPath) with
             {
                 Name = "Malicious Thread",
-                NodeType = ThreadNodeType.NodeType
+                NodeType = MarkdownNodeType.NodeType
             };
 
             Func<Task> act = () => NodeFactory.CreateNode(threadNode).FirstAsync().ToTask();
 
             // Assert
             await act.Should().ThrowAsync<UnauthorizedAccessException>(
-                "User should NOT be able to create threads under another user's node");
+                "User should NOT be able to create nodes under another user's node");
         }
         finally
         {
